Windows fixes.
Reorder member variables so boost::asio objects are properly destroyed.
Clean up integer conversion warnings.
Replace std::chrono with boost::chrono because Windows implementation is
buggy.
Define WIN32_WINNT to be Windows 7

+++ kernel/spline.cpp
@@ -160,7 +160,7 @@ int Spline::insertControlPoint(double xi) {
 		nd[i] = y_x_[i];
 		++i;
 	}
-	int iret = i;
+	int iret = static_cast<int>(i);
 	nx[i] = xi;
 	ny[i] = px.y;
 	nd[i] = px.y_x;
